Reabsorb the Distributions class into the Random class.
